requirements.txt
apache-airflow==2.6.3
black==23.1a1
flake8==7.2.0
flake8-print==5.0.0
flake8-pytest-style==2.1.0
isort==6.0.1
pyspark==3.4.1
psycopg2-binary==2.9.6
celery==5.3.1
flower==2.0.0
pytest==7.4.0
pytest-clarity==1.0.1
pytest-mock==3.14.0
redis==4.6.0
factory_boy==3.3.3
